requirements.txt
pandas==1.5.3
numpy==1.24.4
pyyaml==6.0.1
boto3==1.28.57
matplotlib==3.7.2
seaborn==0.12.2
autogluon.tabular==0.8.2
shap==0.43.0
sagemaker==2.197.0
scikit-learn==1.2.2 